+++ quantify_scheduler/backends/qblox/operation_handling/factory.py
@@ -54,7 +54,7 @@ def _get_acquisition_strategy(
     """Handles the logic for determining the correct acquisition type."""
 
     protocol = operation_info.data["protocol"]
-    if protocol in ("Trace", "SSBIntegrationComplex"):
+    if protocol in ("Trace", "SSBIntegrationComplex", "ThresholdedAcquisition"):
         if protocol == "Trace" and operation_info.data["bin_mode"] == BinMode.APPEND:
             raise ValueError(
                 f"Trace acquisition does not support APPEND bin mode.\n\n"